+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/attributes.go
@@ -23,10 +23,10 @@ const (
 	AWSSystemVal string        = "aws-api"
 )
 
-var servicemap = map[string]AttributeSetter{
-	dynamodb.ServiceID: DynamoDBAttributeSetter,
-	sqs.ServiceID:      SQSAttributeSetter,
-	sns.ServiceID:      SNSAttributeSetter,
+var servicemap = map[string]AttributeBuilder{
+	dynamodb.ServiceID: DynamoDBAttributeBuilder,
+	sqs.ServiceID:      SQSAttributeBuilder,
+	sns.ServiceID:      SNSAttributeBuilder,
 }
 
 // SystemAttr return the AWS RPC system attribute.
@@ -56,11 +56,18 @@ func RequestIDAttr(requestID string) attribute.KeyValue {
 
 // DefaultAttributeSetter checks to see if there are service specific attributes available to set for the AWS service.
 // If there are service specific attributes available then they will be included.
+// Deprecated: Kept for backward compatibility, use DefaultAttributeBuilder instead. This will be removed in a future release.
 func DefaultAttributeSetter(ctx context.Context, in middleware.InitializeInput) []attribute.KeyValue {
+	return DefaultAttributeBuilder(ctx, in, middleware.InitializeOutput{})
+}
+
+// DefaultAttributeBuilder checks to see if there are service specific attributes available to set for the AWS service.
+// If there are service specific attributes available then they will be included.
+func DefaultAttributeBuilder(ctx context.Context, in middleware.InitializeInput, out middleware.InitializeOutput) []attribute.KeyValue {
 	serviceID := v2Middleware.GetServiceID(ctx)
 
 	if fn, ok := servicemap[serviceID]; ok {
-		return fn(ctx, in)
+		return fn(ctx, in, out)
 	}
 
 	return []attribute.KeyValue{}