+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/config.go
@@ -4,6 +4,11 @@
 package otelaws // import "go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-sdk-go-v2/otelaws"
 
 import (
+	"context"
+
+	"github.com/aws/smithy-go/middleware"
+
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 )
@@ -11,7 +16,7 @@ import (
 type config struct {
 	TracerProvider    trace.TracerProvider
 	TextMapPropagator propagation.TextMapPropagator
-	AttributeSetter   []AttributeSetter
+	AttributeBuilders []AttributeBuilder
 }
 
 // Option applies an option value.
@@ -48,9 +53,22 @@ func WithTextMapPropagator(propagator propagation.TextMapPropagator) Option {
 }
 
 // WithAttributeSetter specifies an attribute setter function for setting service specific attributes.
-// If none is specified, the service will be determined by the DefaultAttributeSetter function and the corresponding attributes will be included.
+// If none is specified, the service will be determined by the DefaultAttributeBuilder function and the corresponding attributes will be included.
 func WithAttributeSetter(attributesetters ...AttributeSetter) Option {
+	var attributeBuilders []AttributeBuilder
+	for _, setter := range attributesetters {
+		attributeBuilders = append(attributeBuilders, func(ctx context.Context, in middleware.InitializeInput, out middleware.InitializeOutput) []attribute.KeyValue {
+			return setter(ctx, in)
+		})
+	}
+
+	return WithAttributeBuilder(attributeBuilders...)
+}
+
+// WithAttributeBuilder specifies an attribute setter function for setting service specific attributes.
+// If none is specified, the service will be determined by the DefaultAttributeBuilder function and the corresponding attributes will be included.
+func WithAttributeBuilder(attributeBuilders ...AttributeBuilder) Option {
 	return optionFunc(func(cfg *config) {
-		cfg.AttributeSetter = append(cfg.AttributeSetter, attributesetters...)
+		cfg.AttributeBuilders = append(cfg.AttributeBuilders, attributeBuilders...)
 	})
 }